LayoutTests/inspector/extensions-audits-api.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="extensions-test.js"></script>
<script src="audits-tests.js"></script>
<script src="extensions-audits-tests.js"></script>

<script type="text/javascript">

function extension_testAuditsAPI(nextTest)
{
    function onStartAuditCategory(results)
    {
        log("category.onAuditStarted fired, results dump follows:");
        dumpObject(results);
        var node = results.createResult("Subtree");
        dumpObject(node);
        results.done();
    }
    var category = webInspector.audits.addCategory("Extension audits", 20);
    category.onAuditStarted.addListener(onStartAuditCategory);
    log("Added audit category, result dump follows:");
    dumpObject(category);
    extension_runAudits(nextTest);
}

function runAudits()
{
    evaluateInWebInspector("frontend_runAudits");
}

</script>
</head>
<body onload="onload()">
<p>Tests audits support in WebInspector Extensions API</p>
</body>
</html>
